Factor out automated testing API definitions into monitorlib
