+++ monitoring/rid_qualifier/flight_data_generator.py
@@ -3,7 +3,7 @@ import shapely.geometry
 from pyproj import Geod, Transformer, Proj
 import json
 from pathlib import Path
-from typing import List, Optional
+from typing import List
 import arrow
 import datetime
 from datetime import datetime, timedelta
